+++ src/scenes.jl
@@ -550,7 +550,7 @@ function center!(scene::Scene, padding=0.01, exclude = not_in_data_space)
     bb = transformationmatrix(scene)[] * bb
     w = widths(bb)
     padd = w .* padding
-    bb = Rect3f(minimum(bb) .- padd, w .+ 2padd)
+    bb = Rect3d(minimum(bb) .- padd, w .+ 2padd)
     update_cam!(scene, bb)
     scene
 end
